infra/scripts/ci/hestia_ci/requirements.txt
python-gitlab
pytest